+++ RetroRemakes/Light.h
@@ -12,8 +12,8 @@ public:
 	Light(GLfloat red, GLfloat green, GLfloat blue, GLfloat aIntensity,
 		  GLfloat xDir, GLfloat yDir, GLfloat zDir, GLfloat dIntensity);
 
-	void UseLight(GLfloat ambientIntensityLocation, GLfloat colorLocation,
-				  GLfloat diffuseIntensityLocation, GLfloat directionLocation);
+	void UseLight(GLuint ambientIntensityLocation, GLuint colorLocation,
+				  GLuint diffuseIntensityLocation, GLuint directionLocation);
 
 	~Light();
